refactor(stdio): rename stdio module name
