+++ packages/eval/src/stdlib/mod.rs
@@ -3,7 +3,7 @@ mod len;
 mod map;
 mod nth;
 mod push;
-mod stdout;
+mod stdio;
 
 use std::rc::Rc;
 
@@ -11,13 +11,13 @@ use input::stdlib_input;
 use len::stdlib_len;
 use nth::stdlib_nth;
 use push::stdlib_push;
-use stdout::println::stdlib_println;
+use stdio::println::stdlib_println;
 
 use crate::{
     scope::Scope,
     stdlib::{
         map::{map_get::stdlib_map_get, map_set::stdlib_map_set},
-        stdout::print::stdlib_print,
+        stdio::print::stdlib_print,
     },
     Extension, Plugin,
 };
