+++ packages/eval/src/stdlib/stdio/literal_display.rs
